Added extra operator rule to solver
I believe the solver is missing an operator rule. Consider a board with the next line:

  [1234] < [?] > [1234]

In this case, the possible values for the ? cell are 34 instead of 234, because the left and right cell cannot both contain value 1.
Similarly, a board with the next line (board size = 4):

  [1234] > [?] < [1234]

has possibilities 12 for cell ? instead of 123.

I'm not very familiar with python, so the code in my PR is likely suboptimal but it seems to fix this.